# Conflicts:
#	rust/Cargo.lock
#	rust/crates/tidb-exec/src/real_tikv_stats.rs
#	rust/crates/tidb-executor/Cargo.toml
#	rust/crates/tidb-session/src/dispatch.rs
#	rust/crates/tidb-stats-handle-cache-internal-lfu/src/lib.rs
#	rust/crates/tidb-util/src/cgroup.rs
#	rust/crates/tidb-util/src/memory/mod.rs
#	rust/crates/tidb-util/src/memory/process.rs
#	rust/crates/tidb-util/src/plancodec.rs
#	rust/testport/TESTPORT_EXECPLAN.md
#	rust/testport/receipts/util_breakpoint.md
#	rust/testport/receipts/util_cgmon.md
#	rust/testport/receipts/util_cgroup.md
#	rust/testport/receipts/util_extsort.md
#	rust/testport/receipts/util_gctuner_audit.md
#	rust/testport/receipts/util_memoryusagealarm.md
#	rust/testport/receipts/util_plancodec.md

+++ rust/crates/tidb-dxf/src/task_type.rs
@@ -29,7 +29,7 @@ pub const BACKFILL: TaskType = TaskType::from_static("backfill");
 
 /// Go `Type2Int`: converts a task type to an int.
 #[must_use]
-pub fn type2int(t: &TaskType) -> i64 {
+pub fn type2int(t: &TaskType) -> isize {
     if *t == TASK_TYPE_EXAMPLE {
         1
     } else if *t == IMPORT_INTO {
@@ -43,7 +43,7 @@ pub fn type2int(t: &TaskType) -> i64 {
 
 /// Go `Int2Type`: converts an int to a task type.
 #[must_use]
-pub fn int2type(i: i64) -> TaskType {
+pub fn int2type(i: isize) -> TaskType {
     match i {
         1 => TASK_TYPE_EXAMPLE,
         2 => IMPORT_INTO,
